backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pymongo==4.6.1
dnspython==2.4.2
pyjwt==2.8.0
yt-dlp==2023.11.16
spotipy==2.23.0
requests==2.31.0
werkzeug==3.0.1
ffmpeg-python==0.2.0
